refactor to separate out newsletter-specific components
